Enhanced pan gesture managment.
The panGesture is automatically added to all views managed by NavigationViewController and set to react only for gesture performed on navigation bar. After revealing the menu the gesture area is set to whole view for easy slide back. When the topView fills the whole screen again the gesture is reasingned to its original area. The manuallyAddGestureRecognizerToViewController method allows to add a gesture recognizer to a specific place on the view and changes the value of topViewControllerHasManuallySetGestureRecognizer to true. This BOOL is used by resetGestureToPreviousGestureArea to check if the gesture should be reasigned to view.